[build] fix `Microsoft.NET.Sdk.Android.Manifest-*` version band (#10120)
Backport of: https://github.com/dotnet/android/pull/10112

In fc72bb40, we migrated to a 9.0.300 .NET SDK, which unfortunately
resulted in an Android workload manifest with a 9.0.300 version band:

    Microsoft.NET.Sdk.Android.Manifest-9.0.300.35.0.73.nupkg

To stay on 9.0.100, we need should use `$(DotNetAndroidManifestVersionBand)`
for the manifest version band instead of `$(DotNetSdkManifestsFolder)`.

I also had to make a similar change in `DotNet.targets`.