[CoreCLR] Support for fastdev assemblies (#10113)
Context: https://github.com/dotnet/android/pull/10065
Context: https://github.com/dotnet/android/pull/10075

This is the last of the PR series which implement support for `.NET for Android`
applications Debug builds.

While developing applications using Visual Studio, developers are in the cycle
of making frequent changes and testing them on device or emulator. This development
loop must be fast enough to not make the experience frustrating. One of the ways
`.NET for Android` uses to make it better is the so-called "fastdev" mode in which
the application APK is built and deployed without any assemblies in it, while the
assemblies are synchronized individually to the device's filesystem, so they can
be refreshed individually on changes affecting, thus saving time.

The `.NET for Android` runtime must be able to locate those assemblies and load them
from the filesystem, instead of using the usual assembly store included in the application
package.

This commit implements support for finding, enumerating and locating the fastdev assemblies
on the filesystem. It also modifies one of the "install and run" unit tests to test the
functionality.