[native] Better p/invoke handling when loading foreign libraries (#10165)
Improve looking for "foreign" libraries when resolving a p/invoke symbol.

Foreign libraries are those which either aren't contained in the application APK
or which reside on the device in a system location. In such cases, we must attempt
to load the library using its name passed from the managed code (which uses the
`DllImport` attribute to name the library).

Since `DllImport` doesn't usually use full name of the target library (e.g. instead
of `liblog.so` it will use either `liblog` or `log`), we must modify the library
name so that the `dlopen(3)` call can find it on the filesystem. We must make sure
that the library name follows the `libNAME.so` convention before attempting to load
the library from filesystem..

This is an interim fix for us not finding e.g. the `liblog.so` library when the
`[DllImport("liblog")]` attribute is used in the managed code. The library resolution
and reloading system must be redesigned, which will happen in a separate PR as it is
currently not the highest priority task.