[Mono.Android] JNI handles are now in a "control block" (#10179)
Originally from: https://github.com/dotnet/android/pull/10125
Context: https://github.com/dotnet/java-interop/pull/1339

These are changes to Mono's GC bridge implementation, such that the
same code can be used on both Mono and CoreCLR.

The base `JavaObject` type has changes such as:

```diff
--[NonSerialized] IntPtr                  handle;
--[NonSerialized] JniObjectReferenceType  handle_type;
--[NonSerialized] IntPtr                  weak_handle;
--[NonSerialized] int                     refs_added;
++unsafe  JniObjectReferenceControlBlock* jniObjectReferenceControlBlock;
```

And the new `JniObjectReferenceControlBlock` struct is defined as:

    internal struct JniObjectReferenceControlBlock {
        public	IntPtr  handle;
        public  int     handle_type;
        public  IntPtr  weak_handle;
        public  int     refs_added;
    }

* Each `JavaObject` allocation now has a `NativeMemory.AllocZeroed()`
  call for this struct.

* However, we only have to read a single field now using Mono's native
  embedding API:

```diff
--info->handle            = mono_class_get_field_from_name (info->klass, const_cast<char*> ("handle"));
--info->handle_type       = mono_class_get_field_from_name (info->klass, const_cast<char*> ("handle_type"));
--info->refs_added        = mono_class_get_field_from_name (info->klass, const_cast<char*> ("refs_added"));
--info->weak_handle       = mono_class_get_field_from_name (info->klass, const_cast<char*> ("weak_handle"));
++info->jniObjectReferenceControlBlock    = mono_class_get_field_from_name (info->klass, const_cast<char*> ("jniObjectReferenceControlBlock"));
```

We are hoping this results in ~the same performance as before, with
the flexibility to support multiple GC bridge implementations and
runtimes.

Other changes:

* Fix SIGSEGV when using the new control block

We were using it incorrectly - there's no need to fetch/set values of managed object fields.
The correct thing to do is to read/write them on the control block itself.

Co-authored-by: Marek Habersack <grendel@twistedcode.net>