[CoreCLR] Compressed assemblies code producess less relocs
Instead of separate per-assembly buffers, there's now a single buffer
and assembly descriptors use offsets into this buffer to store data.
This removes N + 1 relocs where N is the number of assemblies in the app.
The extra removed relocation is for the pointer which used to point to
the descriptors array, which is now made public.

MonoVM runtime not yet updated.