[xabt] remove `$(AndroidStripILAfterAOT)` (#10204)
Fixes: https://github.com/dotnet/android/issues/9577

Partially reverts 82a40927.

We introduced an "experimental" property `$(AndroidStripILAfterAOT)`
in .NET 8, which has a few flaws as mentioned in #9577. I've seen this
come up on a couple customer issues, and the solution was to stop
using the feature...

Since, this feature is only available for Mono (not CoreCLR or
NativeAOT), it feels like we should remove it from the codebase for
now. We could bring it back in the future, if we also solved #9577.