[CoreCLR] Remove unnecessary logging (#10247)
Logging to logcat in Android is expensive and unpredictable in
terms of how much time it takes exactly.  Therefore, we try to
log as little as possible by default, reducing the delays.

This PR removes (or demotes to a lower priority) a number of log 
messages that were useful during testing and development, but are
an unnecessary noise for day-to-day app operation.