Return empty properties when `java` fails to run (#255)
If the Java executable we are trying to run fails to be be executed
(i.e., locked file, broken symlink, or any raeson actually) we should
just return an empty dictionary. That way when running
`GetPathEnvironmentJdkPaths`, the invalid/not available executable is
just ignored.

Call stack from:

* https://developercommunity.visualstudio.com/t/visual-studio-does-not-see-detect-the-co/10888076

    System.ComponentModel.Win32Exception (0x80004005): The system cannot find the path specified
       at System.Diagnostics.Process.StartWithCreateProcess(ProcessStartInfo startInfo)
       at Xamarin.Android.Tools.ProcessUtils.Exec(ProcessStartInfo processStartInfo, DataReceivedEventHandler output, Boolean includeStderr)
       at Xamarin.Android.Tools.JdkInfo.GetJavaProperties(Action`2 logger, String java)
       at Xamarin.Android.Tools.JdkInfo.<GetPathEnvironmentJdkPaths>d__61.MoveNext()

The message `The system cannot find the path specified` is a bit
misleading because we are already validating the file exist before
calling `GetJavaProperties`, and this can only occurr if we found a
broken symlink.