[CoreCLR] Generate less relocations in the native typemap code (#10225)
Context: 96a73e68ed4a6c8f6c3dc683547605c5be5edaa7
Context: f37158eae0111a6cdd25e870e412fcc74997d529

This is another PR in the series that remove native relocations from
the `libxamarin-app.so` library. 

Modify the typemap data so that no pointers to per-module data for
the primary and duplicate maps need to be stored. Module here refers
to a single assembly used by the application. Each assembly/module is
described by an instance of a C++ `struct` which used to contain up to
two pointers to tables containing actual type mapping data - one for the
primary map (all modules have this one) and an optional map of duplicate
types, which was stored in another pointer.

The two pointers pointed to separate arrays of map data structures. Instead,
this PR now puts all of the primary map data for all modules in one array,
and all of the duplicate map data in another array. The assembly/module descriptor
structures now merely contain indices into those arrays.

This removes up to X*2 relocations (with X being the number of assemblies).

The change gives us another modest startup performance boost of 0.67% on a Pixel 8
device, using the `dotnet new maui -sc` sample. It also decreases the number of
relocations in `libxamarin-app.so` from 394 to 70 entries for this sample app.

Modest size decrease of `libxamarin-app.so` is another outcome:

  * arm64: 370936 to 363976 bytes (1.87%)
  * x64: 362520 to 355496 bytes (1.93%)