[CoreCLR] Ignore some assemblies when asked to load them (#10249)
Ignore assemblies whose names end in `.ni` and `.ni.dll`, when 
a load request is made via the CoreCLR host contract callback.

For each CoreCLR will ask to first load `Assembly.ni.dll`, NI
standing for Native Image. This is something that was used on
Windows long ago and it never was, nor will it be, supported on
Unix.

This is mostly a cosmetic change in that it prevents a false
warning about a missing assembly from being logged in logcat.
This, in itself, is a startup time improvement since those
warnings would always be logged during every application launch,
with logcat logging being very expensive (and unpredictable).

The message about a missing `.ni.dll` is still logged, but with
the Debug severity, which will cause it to show up only if the
`assembly` logging category is enabled (it is disabled by default)