[xabt] fix missing `.nupkg` output for `dotnet pack` (#10270)
Fixes: https://github.com/dotnet/android/issues/9974

After updating the `DotnetPack` test to use:

    proj.OtherBuildItems.Add (new AndroidItem.AndroidMavenLibrary ("org.jetbrains.kotlinx:kotlinx-serialization-json-jvm") {
        MetadataValues = "Version=1.3.3;Bind=false",
        BinaryContent = () => [],
    });

And then seeing the `.jar` files inside:

    using var aarStream = new MemoryStream ();
    var aarEntry = nupkg.ReadEntry (aarPath);
    aarEntry.Extract (aarStream);
    aarStream.Seek (0, SeekOrigin.Begin);
    using var aar = ZipArchive.Open (aarStream);
    int count = aar.Count (e =>
        e.FullName.StartsWith ("libs/", StringComparison.OrdinalIgnoreCase) &&
        e.FullName.EndsWith (".jar", StringComparison.OrdinalIgnoreCase));

There were originally 0 files, when there *should* be 2!

Reviewing the `.binlog`, `dotnet pack` has an inner/outer build going on during it:

* `_WalkEachTargetPerFramework` target

* Runs `<MSBuild/>`, the `_GetFrameworkAssemblyReferences` target

* Which calls our `_CreateAar` target.

If `_CreateAar` can be called in this scenario, it needs to depend on `_CategorizeAndroidLibraries`, so none of its inputs are missing.

With this change, the test now asserts that 2 "hashed" `.jar` files are present in the `.nupkg` output.