[native] Ensure `function_name()` is part of log message (#10302)
Context: https://github.com/dotnet/runtime/issues/117733

When a .NET for Android app aborts, the abort message should contain
the file name, line number, column, and function name of the function
which triggered the abort.

Insteda, we get the filename, line number, and column, but we get
`%s` for the function name:

    F monodroid: Abort at mono-log-adapter.cc:46:3 ('%s')

Update the `log_fatal()` invocation within `mono-log-adapter.cc`
to use `{}` instead of `%s`, in accordance with C++ `std::format`.
This should cause `sloc.function_name()` to be properly part of the
abort message.