[tests] fix `GetString_Many()` test
The test is currently failing with:

    Expected: True
    But was:  False
    Stack trace
       at Xamarin.Android.NetTests.HttpClientIntegrationTestBase.GetString_Many()
       at System.Reflection.MethodBaseInvoker.InterpretedInvoke_Method(Object obj, IntPtr* args)
       at System.Reflection.MethodBaseInvoker.InvokeWithNoArgs(Object , BindingFlags )

If I navigate to http://example.org in a browser, I just get a DNS error.

There are other tests in here that use https://google.com instead, so
I switched to use it.

I also removed the `MobileNotWorking` category, which doesn't appear
to be actually ignored because the test is running?