[build] set `$(PackageVersionSuffix)` for dotnet/android-tools (#10325)
Context: https://github.com/dotnet/android-tools/pull/259

This adds the suffix:

    <PackageVersionSuffix>-preview.$(PackVersionCommitCount)</PackageVersionSuffix>

In order to allow the package to be published as a "transport" package
with a non-stable version:

    packages\microsoft.dotnet.arcade.sdk\10.0.0-beta.25367.5\tools\SdkTasks\PublishArtifactsInManifest.proj(124,5): error :
    Package 'Xamarin.Android.Tools.AndroidSdk' has stable version '1.0.104' but is targeted at a non-isolated feed 'https://pkgs.dev.azure.com/dnceng/public/_packaging/dotnet10/nuget/v3/index.json'

I also set this logic to only run when `$(IsPacking)` is true, as
there are build ordering problems during `make prepare` without
doing this. The bootstrap tasks `.dll`'s don't exist then.