[build] remove usage of "Facades" directories
Context: https://github.com/dotnet/android/pull/10339

In order to complete #10339, I want to refactor & remove the need for
having to define `$(_XATargetFrameworkDirectories)` at all.

An easy first step is to remove the usage of "Facades" directories
because these don't exist in the .NET 6+ world. These are a Mono/.NET
framework-ism, which is replaced by "ref" packs in .NET 6+.

I also removed the `gen-ProfileAssemblies.sh` script, which was used to
generate files specific to Mono & Xamarin.Android, but is no longer
needed.