[copilot] update instructions file (#10355)
There is a new Copilot > Configure Chat... > Generate Instructions menu in VS Code.

I let it run and update the file, and accepted the parts that looked useful.

The most important thing, it found various links to `.md` files that were broken!

This makes the file about 50% smaller, which may give @copilot more context window to work with.