[native] Use full marshal method signature when a method is overloaded
Fixes: #10417
Context: https://docs.oracle.com/en/java/javase/17/docs/specs/jni/design.html#resolving-native-method-names
Context: https://github.com/dotnet/android/issues/10417#issuecomment-3210789627

Whenever we generate marshal methods code for overloaded Java methods, we must be sure to use full native signature
instead of the short form, even if one of the overloads doesn't have parameters.

Failure to do so results in hard-to fix subtle bugs like in case of #10417. In this instance, we had a
class (`Android.Runtime.InputStreamAdapter`) with 3 method overloads:

  * `public override int Read ();`
  * `public override int Read (byte[] bytes)`
  * `public override int Read (byte[] bytes, int offset, int length)`

The sample application failed to load an image using a Glide decoder, which called the `Read (byte[], int, int)`
overload above. It was puzzling, because the Java code clearly called the right native function and yet, it ended up in
a marshal method wrapper for the `Read ()` overload.

The cause became apparent when I checked names of the generated native functions. All of them were valid and yet the
whole was subtly broken: the `Read ()` overload used a (valid) "short form" native symbol name, which omits the part
that encodes parameter types in the symbol name. The reason for this was that the generator code saw there were no
parameters and so it output the short form, because it is the **first** symbol name looked up by JavaVM - thus it seemed
to be a good optimization at the time of implementing the marshal methods generator.

The problems began when **both** overloads were used somewhere in the application code and the `Read()` overload
was **always** found first, since it used the short form name, and thus Java never got to looking up the other overload!
Then it proceeded to call the parameter-less overload, which caused sort-of worked (as in it didn't crash) it it was far
from correct.

The fix implemented here is to always emit the `__` native symbol suffix for any overloaded methods that have no
parameters.