[native] Make native runtime smaller and allow optional .so stripping (#10446)
Applications built for .NET10 gained some size because of a number of factors. One of
them is our decision earlier this year to start statically linking `libc++` shipped with
the NDK. The reason for this was a desire to start using `std::format`, which provides
a fully type-safe and it parses format strings at the compilation time, making run time
faster. This added around 200k to the runtime.

However, another side effect of that decision is that the ELF sections used by exception
handling got much bigger and since we don't use exceptions, those sections are a ballast
and nothing else.

This PR does two things:

  * it removes 3 exception handling sections from the runtimes
    (`eh_frame`, `eh_frame_hdr` and `.gcc_except_table`) saving us around 120k
  * it adds an option to strip all native libraries when packaging the application.

By default we produce shared libraries with the symbol table present, to make potential
native crash stack traces easier to read and interpret. This, however, adds around 400k
to our runtime, so it might not be desirable for everybody to keep that baggage around.

Introduce the `AndroidStripNativeLibraries` boolean MSBuild property, with a default
value of `false` which, if set to `true`, will cause all the native shared libraries
to be stripped of debug symbols at packaging time.